fix:それぞれのロールによって、playviewに表示する内容を変更できるように
