+++ kei-doro/Assets.xcassets/MainColor.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "1.000",
-          "green" : "1.000",
-          "red" : "0.136"
+          "blue" : "0xFF",
+          "green" : "0xFF",
+          "red" : "0x22"
         }
       },
       "idiom" : "universal"
